requirements.txt
llama-cpp-python==0.1.85
huggingface_hub==0.23.2
pandas==1.5.3
tiktoken==0.6.0
pymupdf==1.25.1
langchain==0.1.1
langchain-community==0.0.13
chromadb==0.4.22
sentence-transformers==2.3.1
numpy==1.25.2